requirement.txt
python-dotenv
pandas
numpy
pymongo==4.13.0
pymongo[srv]
certifi
scikit-learn
pyaml
dill

##-e .